ci: locked board deletion for admins.
- tweaked deleteSavedBoard redis method to delete all board ids of all user sets.
- added admin avatar on boards that are shared with a user.
- added new trpc router for board users.

+++ src/server/routers/board-router/routes/delete-board.ts
@@ -1,9 +1,13 @@
-import { internalServerError } from '@server/helpers/error-helpers';
+import {
+    internalServerError,
+    unauthorized,
+} from '@server/helpers/error-helpers';
 import { deleteSavedBoard } from '@server/redis/board';
 import { authedProcedure } from '@server/routers/auth-router';
 import { z } from 'zod';
 import { BOARD_CACHE_DELETE_ERROR } from '../errors';
 import { deleteError } from '@server/routers/common-errors';
+import { queryBoardUserRole } from '@server/routers/board-user-router/routes/get-board-user-role';
 
 export const deleteBoard = authedProcedure
     .input(
@@ -12,18 +16,34 @@ export const deleteBoard = authedProcedure
         })
     )
     .mutation(async ({ ctx, input }) => {
+        const id = input.id;
+        const userId = ctx.session.user.id;
+
+        const role = await queryBoardUserRole(userId, id, ctx.prisma);
+        if (role !== 'ADMIN') {
+            throw unauthorized('Your user is not an admin of the board.');
+        }
+
         try {
             const board = await ctx.prisma.board.delete({
                 where: {
-                    id: input.id,
+                    id,
+                },
+                include: {
+                    boardUser: {
+                        select: {
+                            userId: true,
+                        },
+                    },
                 },
             });
 
-            await deleteSavedBoard(ctx.session.user.id, board.id).catch(
-                (error) => {
-                    throw internalServerError(BOARD_CACHE_DELETE_ERROR, error);
-                }
-            );
+            await deleteSavedBoard(
+                board.boardUser.map((user) => user.userId),
+                board.id
+            ).catch((error) => {
+                throw internalServerError(BOARD_CACHE_DELETE_ERROR, error);
+            });
 
             return board;
         } catch (error) {
